fix: Move botdetector code into separate file
